backend/requirements.txt
asgiref==3.9.2
Django==5.2.6
djangorestframework==3.16.1
python-dotenv==1.1.1
sqlparse==0.5.3
pytest==8.3.2
pytest-django==4.9.0
responses==0.25.3
cryptography==43.0.1
djangorestframework-simplejwt==5.3.1
